Rename module, change code in message.py to reflect change
